% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISP_functions.R
\name{plot.sim.data}
\alias{plot.sim.data}
\title{Plot Mean Model for Data.}
\usage{
\method{plot}{sim.data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sim.data}, which results from running the \code{\link{sim.data}} function.}

\item{...}{Additional arguments to be passed, which are ignored in this function.}
}
\value{
None.
}
\description{
This function plots the mean model for the scenario from which data was generated using \code{\link{sim.data}}.
}
\examples{
#See ?'crisp-package' for a full example of how to use this package

#generate data (using a very small 'n' for illustration purposes)
set.seed(1)
data <- sim.data(n = 15, scenario = 2)

#plot the mean model for the scenario from which we generated data
plot(data)
}
\seealso{
\code{\link{sim.data}}
}

